refactored posts to work with root nesting
